Increase robustness to exotic clang versions
As seen in #162, not all clang installations neatly follow the "clang
version MAJOR.MINOR" template. This causes crofiler to fail when using,
e.g., Debian's provided version of clang. This commit increases the
robustness of the version checking by relying on a regular expression
which can match a broader range of possible versions.

Resolves #162.